feat: CustomerRepository 인터페이스 생성 및 기존 CustomerRepository 이름 변경
